explicit cast to uint64_t

+++ include/mc_cpp/anvil/anvil_chunk.hpp
@@ -86,7 +86,7 @@ namespace mc::anvil {
     template<typename SectionDataType>
     [[nodiscard]]
     static uint64_t getBitsPerIndex(const size_t length) {
-        return std::max<uint64_t>(std::bit_width(std::max(length, 1UL) - 1), SectionDataInfo<SectionDataType>::paletteMinBits);
+        return std::max<uint64_t>(std::bit_width(std::max(static_cast<uint64_t>(length), static_cast<uint64_t>(1)) - 1), SectionDataInfo<SectionDataType>::paletteMinBits);
     }
 
     template<typename SectionDataType>
